substrabac/requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.3.0
Django==2.0.5
django-filter==1.1.0
djangorestframework==3.8.2
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
Markdown==2.6.11
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2018.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
